chore: rename only non shadcn ones
