+++ app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/useAssignTicket.ts
@@ -1,7 +1,7 @@
 import { useConversationContext } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/conversationContext";
 import { useConversationListContext } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/list/conversationListContext";
 import { useConversationsListInput } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/shared/queries";
-import { toast } from "@/components/hooks/useToast";
+import { toast } from "@/components/hooks/use-toast";
 import { api } from "@/trpc/react";
 
 export const useAssignTicket = () => {
