+++ app/(dashboard)/mailboxes/[mailbox_slug]/[category]/inbox.tsx
@@ -14,7 +14,7 @@ import {
 import { TabBar } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/tabBar";
 import { useSaveLatestMailboxSlug } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/useSaveLatestMailboxSlug";
 import { FileUploadProvider } from "@/components/fileUploadContext";
-import { useIsMobile } from "@/components/hooks/useMobile";
+import { useIsMobile } from "@/components/hooks/use-mobile";
 import LoadingSpinner from "@/components/loadingSpinner";
 import { PageHeader } from "@/components/pageHeader";
 import useKeyboardShortcut from "@/components/useKeyboardShortcut";
