+++ app/(dashboard)/mailboxes/[mailbox_slug]/settings/customers/customerSetting.tsx
@@ -1,13 +1,13 @@
 "use client";
 
 import { useState } from "react";
-import { useSavingIndicator } from "@/components/hooks/useSavingIndicator";
-import { toast } from "@/components/hooks/useToast";
+import { toast } from "@/components/hooks/use-toast";
 import { SavingIndicator } from "@/components/savingIndicator";
 import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Separator } from "@/components/ui/separator";
+import { useSavingIndicator } from "@/components/ui/useSavingIndicator";
 import { useDebouncedCallback } from "@/components/useDebouncedCallback";
 import { useOnChange } from "@/components/useOnChange";
 import { RouterOutputs } from "@/trpc";
