+++ app/(dashboard)/mailboxes/[mailbox_slug]/settings/knowledge/websiteCrawlSetting.tsx
@@ -5,7 +5,7 @@ import { Clock, PlusCircle, RefreshCw, Trash } from "lucide-react";
 import { useParams } from "next/navigation";
 import { useState } from "react";
 import { ConfirmationDialog } from "@/components/confirmationDialog";
-import { toast } from "@/components/hooks/useToast";
+import { toast } from "@/components/hooks/use-toast";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
