+++ app/(dashboard)/mailboxes/[mailbox_slug]/settings/preferences/mailboxNameSetting.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import { useState } from "react";
-import { useSavingIndicator } from "@/components/hooks/useSavingIndicator";
-import { toast } from "@/components/hooks/useToast";
+import { toast } from "@/components/hooks/use-toast";
 import { SavingIndicator } from "@/components/savingIndicator";
 import { Input } from "@/components/ui/input";
+import { useSavingIndicator } from "@/components/ui/useSavingIndicator";
 import { useDebouncedCallback } from "@/components/useDebouncedCallback";
 import { useOnChange } from "@/components/useOnChange";
 import { RouterOutputs } from "@/trpc";
