+++ app/(dashboard)/mailboxes/[mailbox_slug]/settings/tools/apiCard.tsx
@@ -3,7 +3,7 @@
 import { Check, RefreshCw, Trash } from "lucide-react";
 import { useState } from "react";
 import { ConfirmationDialog } from "@/components/confirmationDialog";
-import { toast } from "@/components/hooks/useToast";
+import { toast } from "@/components/hooks/use-toast";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Label } from "@/components/ui/label";
