+++ components/savingIndicator.tsx
@@ -1,5 +1,5 @@
 import { Check, Loader2, X } from "lucide-react";
-import type { SavingState } from "@/components/hooks/useSavingIndicator";
+import type { SavingState } from "@/components/ui/useSavingIndicator";
 import { cn } from "@/lib/utils";
 
 interface SavingIndicatorProps {
