+++ components/useSpeechRecognition.ts
