+++ tests/components/hooks/useSavingIndicator.test.ts
@@ -3,7 +3,7 @@
  */
 import { act, renderHook } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { useSavingIndicator } from "@/components/hooks/useSavingIndicator";
+import { useSavingIndicator } from "@/components/ui/useSavingIndicator";
 
 describe("useSavingIndicator", () => {
   beforeEach(() => {
